@@ -24,7 +24,7 @@ namespace GeoJSON.Net.Converters
         /// </returns>
         public override bool CanConvert(Type objectType)
         {
-            return typeof(IGeometryObject).IsAssignableFromType(objectType);
+            return typeof(IGeoJSONObject).IsAssignableFromType(objectType);
         }
 
         /// <summary>
@@ -48,7 +48,7 @@ namespace GeoJSON.Net.Converters
                     return ReadGeoJson(value);
                 case JsonToken.StartArray:
                     var values = JArray.Load(reader);
-                    var geometries = new ReadOnlyCollection<IGeometryObject>(
+                    var geometries = new ReadOnlyCollection<IGeoJSONObject>(
                         values.Cast<JObject>().Select(ReadGeoJson).ToArray());
                     return geometries;
             }
@@ -80,49 +80,36 @@ namespace GeoJSON.Net.Converters
         /// <exception cref="System.NotSupportedException">
         /// Feature and FeatureCollection types are Feature objects and not Geometry objects
         /// </exception>
-        private static IGeometryObject ReadGeoJson(JObject value)
+        private static IGeoJSONObject ReadGeoJson(JObject value)
         {
-            JToken token;
-
-            if (!value.TryGetValue("type", StringComparison.OrdinalIgnoreCase, out token))
+            if (!value.TryGetValue("type", StringComparison.OrdinalIgnoreCase, out var token))
             {
                 throw new JsonReaderException("json must contain a \"type\" property");
             }
 
-            GeoJSONObjectType geoJsonType;
 
-#if (NET35)
-            try
-            {
-                geoJsonType = (GeoJSONObjectType)Enum.Parse(typeof(GeoJSONObjectType), token.Value<string>(), true);
-            }
-            catch (Exception)
-            {
-                throw new JsonReaderException("Type must be a valid geojson object type");
-            }
-#else
-            if (!Enum.TryParse(token.Value<string>(), true, out geoJsonType))
+            if (!Enum.TryParse(token.Value<string>(), true, out GeoJSONObjectType geoJsonType))
             {
                 throw new JsonReaderException("type must be a valid geojson geometry object type");
             }
-#endif
+
 
             switch (geoJsonType)
             {
                 case GeoJSONObjectType.Point:
                     return value.ToObject<Point>();
-                case GeoJSONObjectType.MultiPoint:
-                    return value.ToObject<MultiPoint>();
-                case GeoJSONObjectType.LineString:
-                    return value.ToObject<LineString>();
-                case GeoJSONObjectType.MultiLineString:
-                    return value.ToObject<MultiLineString>();
-                case GeoJSONObjectType.Polygon:
-                    return value.ToObject<Polygon>();
-                case GeoJSONObjectType.MultiPolygon:
-                    return value.ToObject<MultiPolygon>();
-                case GeoJSONObjectType.GeometryCollection:
-                    return value.ToObject<GeometryCollection>();
+                //case GeoJSONObjectType.MultiPoint:
+                //    return value.ToObject<MultiPoint>();
+                //case GeoJSONObjectType.LineString:
+                //    return value.ToObject<LineString>();
+                //case GeoJSONObjectType.MultiLineString:
+                //    return value.ToObject<MultiLineString>();
+                //case GeoJSONObjectType.Polygon:
+                //    return value.ToObject<Polygon>();
+                //case GeoJSONObjectType.MultiPolygon:
+                //    return value.ToObject<MultiPolygon>();
+                //case GeoJSONObjectType.GeometryCollection:
+                //    return value.ToObject<GeometryCollection>();
                 case GeoJSONObjectType.Feature:
                 case GeoJSONObjectType.FeatureCollection:
                 default:
