@@ -1,4 +1,4 @@
-﻿// Copyright © Joerg Battermann 2014, Matt Hunt 2017
+﻿// Copyright © Matt Hunt 2017
 
 using System;
 using System.Collections.Generic;
@@ -31,10 +31,9 @@ namespace GeoJSON.Net.CoordinateReferenceSystem
                 throw new ArgumentNullException(nameof(href));
             }
 
-            Uri uri;
-            if (href.Length == 0 || !Uri.TryCreate(href, UriKind.RelativeOrAbsolute, out uri))
+            if (href.Length == 0 || !Uri.TryCreate(href, UriKind.RelativeOrAbsolute, out _))
             {
-                throw new ArgumentException("must be a dereferenceable URI", nameof(href));
+                throw new ArgumentException("Must be a dereferenceable URI.", nameof(href));
             }
 
             Properties = new Dictionary<string, object> { { "href", href } };
