@@ -1,4 +1,4 @@
-﻿// Copyright © Joerg Battermann 2014, Matt Hunt 2017
+﻿// Copyright © Matt Hunt 2017
 
 using System;
 using System.Collections.Generic;
@@ -11,7 +11,7 @@ namespace GeoJSON.Net
     /// <remarks>
     ///     10 decimal places equates to accuracy to 11.1 μm.
     /// </remarks>
-    public class DoubleTenDecimalPlaceComparer : IEqualityComparer<double>
+    internal class DoubleTenDecimalPlaceComparer : IEqualityComparer<double>
     {
         public bool Equals(double x, double y)
         {
